(HP-2023) Renamed reportGraphQlErrors to reportErrorsToSentry
The default exported function has the same name
