+++ src/common/sentry/reportErrorsToSentry.ts
