+++ src/gdprApi/actions/deleteProfile.ts
@@ -24,7 +24,7 @@ import parseDeleteProfileResult, {
   DeleteResultLists,
 } from '../../profile/helpers/parseDeleteProfileResult';
 import { convertStringToTranslationLanguage } from '../../profile/helpers/createServiceConnectionsQueryVariables';
-import reportErrorsToSentry from '../../common/sentry/reportGraphQlErrors';
+import reportErrorsToSentry from '../../common/sentry/reportErrorsToSentry';
 
 const DELETE_PROFILE = loader('../graphql/GdprDeleteMyProfileMutation.graphql');
